config/validate: move astnode into its own package
This is needed to prevent cyclic dependencies for tests.
