+++ config/validate/validate.go
@@ -20,6 +20,7 @@ import (
 	"reflect"
 	"strings"
 
+	"github.com/coreos/ignition/config/validate/astnode"
 	"github.com/coreos/ignition/config/validate/report"
 )
 
@@ -27,36 +28,10 @@ type validator interface {
 	Validate() report.Report
 }
 
-// AstNode abstracts the differences between yaml and json nodes, providing a
-// common interface
-type AstNode interface {
-	// ValueLineCol returns the line, column, and highlight string of the value of
-	// this node in the source.
-	ValueLineCol(source io.ReadSeeker) (int, int, string)
-
-	// KeyLineCol returns the line, column, and highlight string of the key for the
-	// value of this node in the source.
-	KeyLineCol(source io.ReadSeeker) (int, int, string)
-
-	// LiteralValue returns the value of this node.
-	LiteralValue() interface{}
-
-	// SliceChild returns the child node at the index specified. If this node is not
-	// a slice node, an empty AstNode and false is returned.
-	SliceChild(index int) (AstNode, bool)
-
-	// KeyValueMap returns a map of keys and values. If this node is not a mapping
-	// node, nil and false are returned.
-	KeyValueMap() (map[string]AstNode, bool)
-
-	// Tag returns the struct tag used in the config structure used to unmarshal.
-	Tag() string
-}
-
 // Validate walks down a struct tree calling Validate on every node that implements it, building
 // A report of all the errors, warnings, info, and deprecations it encounters. If checkUnusedKeys
 // is true, Validate will generate warnings for unused keys in the ast, otherwise it will not.
-func Validate(vObj reflect.Value, ast AstNode, source io.ReadSeeker, checkUnusedKeys bool) (r report.Report) {
+func Validate(vObj reflect.Value, ast astnode.AstNode, source io.ReadSeeker, checkUnusedKeys bool) (r report.Report) {
 	if !vObj.IsValid() {
 		return
 	}
@@ -138,11 +113,11 @@ func getFields(vObj reflect.Value) []field {
 	return ret
 }
 
-func validateStruct(vObj reflect.Value, ast AstNode, source io.ReadSeeker, checkUnusedKeys bool) report.Report {
+func validateStruct(vObj reflect.Value, ast astnode.AstNode, source io.ReadSeeker, checkUnusedKeys bool) report.Report {
 	r := report.Report{}
 
 	// isFromObject will be true if this struct was unmarshalled from a JSON object.
-	keys, isFromObject := map[string]AstNode{}, false
+	keys, isFromObject := map[string]astnode.AstNode{}, false
 	if ast != nil {
 		keys, isFromObject = ast.KeyValueMap()
 	}
@@ -154,8 +129,8 @@ func validateStruct(vObj reflect.Value, ast AstNode, source io.ReadSeeker, check
 	tags := []string{}
 
 	for _, f := range getFields(vObj) {
-		// Default to nil AstNode if the field's corrosponding node cannot be found.
-		var sub_node AstNode
+		// Default to nil astnode.AstNode if the field's corrosponding node cannot be found.
+		var sub_node astnode.AstNode
 		// Default to passing a nil source if the field's corrosponding node cannot be found.
 		// This ensures the line numbers reported from all sub-structs are 0 and will be changed by AddPosition
 		var src io.ReadSeeker
